# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your Flask API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Weather"
    },
    {
      "name": "Destinations"
    },
    {
      "name": "Itineraries"
    },
    {
      "name": "Expenses"
    }
  ],
  "paths": {
    "/weather": {
      "get": {
        "summary": "Get weather data",
        "tags": [
          "Weather"
        ],
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "description": "The location for weather data",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "condition": {
                      "type": "string"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "wind_speed": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Weather data not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An error occurred while fetching weather data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/destinations": {
      "post": {
        "summary": "Create a new destination",
        "tags": [
          "Destinations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Destination added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all destinations",
        "tags": [
          "Destinations"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Destination"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/destinations/{destination_id}": {
      "get": {
        "summary": "Get a specific destination",
        "tags": [
          "Destinations"
        ],
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            }
          },
          "404": {
            "description": "Destination not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific destination",
        "tags": [
          "Destinations"
        ],
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Destination updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific destination",
        "tags": [
          "Destinations"
        ],
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destination deleted successfully"
          }
        }
      }
    },
    "/itineraries": {
      "post": {
        "summary": "Create a new itinerary",
        "tags": [
          "Itineraries"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination_id": {
                    "type": "integer"
                  },
                  "activity": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Itinerary added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/itineraries/destination/{destination_id}": {
      "get": {
        "summary": "Get all itineraries for a specific destination",
        "tags": [
          "Itineraries"
        ],
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Itinerary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/itineraries/{itinerary_id}": {
      "get": {
        "summary": "Get a specific itinerary",
        "tags": [
          "Itineraries"
        ],
        "parameters": [
          {
            "name": "itinerary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Itinerary"
                }
              }
            }
          },
          "404": {
            "description": "Itinerary not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific itinerary",
        "tags": [
          "Itineraries"
        ],
        "parameters": [
          {
            "name": "itinerary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination_id": {
                    "type": "integer"
                  },
                  "activity": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Itinerary updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific itinerary",
        "tags": [
          "Itineraries"
        ],
        "parameters": [
          {
            "name": "itinerary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Itinerary deleted successfully"
          }
        }
      }
    },
    "/expenses": {
      "post": {
        "summary": "Create a new expense",
        "tags": [
          "Expenses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination_id": {
                    "type": "integer"
                  },
                  "expense_category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses/destination/{destination_id}": {
      "get": {
        "summary": "Get all expenses for a specific destination",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses/{expense_id}": {
      "get": {
        "summary": "Get a specific expense",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific expense",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "destination_id": {
                    "type": "integer"
                  },
                  "expense_category": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific expense",
        "tags": [
          "Expenses"
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense deleted successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Destination": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "Itinerary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "destination_id": {
            "type": "integer"
          },
          "activity": {
            "type": "string"
          }
        }
      },
      "Expense": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "destination_id": {
            "type": "integer"
          },
          "expense_category": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      }
    }
  }
}